server/requirements.txt
fastapi
fastmcp
httpx
python-dotenv
pytest
py-key-value-aio[redis]
pytest-asyncio
sqlalchemy[asyncio]
asyncpg
alembic
aiosqlite
pydantic
uvicorn
factory_boy
python-jose[cryptography]
tweepy
linkedin-api-client
